+++ src/UiCollapsible.vue
@@ -21,23 +21,25 @@
                 </svg>
             </ui-icon>
 
-            <ui-ripple-ink v-if="!disableRipple && !disabled && isReady"></ui-ripple-ink>
+            <ui-ripple-ink v-if="!disableRipple && !disabled"></ui-ripple-ink>
         </div>
 
         <transition
-            name="ui-collapsible--transition-toggle"
-            @after-enter="onEnter"
-            @after-leave="onLeave"
+            @enter="onEnter"
+            @after-enter="afterEnter"
+            @before-leave="beforeLeave"
+            @leave="onLeave"
         >
             <div
                 class="ui-collapsible__body-wrapper"
-                ref="body"
+                ref="bodyWrapper"
 
-                :style="{ 'height': calculatedHeight }"
+                :aria-hidden="isOpen ? null : 'true'"
+                :id="id"
 
                 v-show="isOpen"
             >
-                <div class="ui-collapsible__body" :aria-hidden="isOpen ? null : 'true'" :id="id">
+                <div class="ui-collapsible__body">
                     <slot></slot>
                 </div>
             </div>
@@ -48,8 +50,6 @@
 <script>
 import UiIcon from './UiIcon.vue';
 import UiRippleInk from './UiRippleInk.vue';
-
-import RespondsToWindowResize from './mixins/RespondsToWindowResize.js';
 import UUID from './helpers/uuid';
 
 export default {
@@ -77,10 +77,7 @@ export default {
 
     data() {
         return {
-            height: 0,
-            isReady: false,
             isOpen: this.open,
-            useInitialHeight: false,
             id: UUID.short('ui-collapsible-')
         };
     },
@@ -91,10 +88,6 @@ export default {
                 { 'is-open': this.isOpen },
                 { 'is-disabled': this.disabled }
             ];
-        },
-
-        calculatedHeight() {
-            return (this.height === 0 || this.useInitialHeight) ? 'initial' : this.height + 'px';
         }
     },
 
@@ -107,24 +100,13 @@ export default {
     },
 
     mounted() {
-        this.isReady = true;
-        this.refreshHeight();
-
-        this.$on('window-resize', () => {
-            this.refreshHeight();
-        });
+        // Remove the max-height to allow the element to grow if it's open initially
+        if (this.isOpen) {
+            this.$refs.bodyWrapper.style.maxHeight = 'none';
+        }
     },
 
     methods: {
-        onEnter() {
-            this.$emit('open');
-            this.refreshHeight();
-        },
-
-        onLeave() {
-            this.$emit('close');
-        },
-
         toggleCollapsible() {
             if (this.disabled) {
                 return;
@@ -133,31 +115,34 @@ export default {
             this.isOpen = !this.isOpen;
         },
 
-        refreshHeight() {
-            const body = this.$refs.body;
+        onEnter(el) {
+            el.style.maxHeight = el.scrollHeight + 'px';
+            this.$emit('open');
+        },
 
-            this.useInitialHeight = true;
-            body.style.display = 'block';
+        afterEnter(el) {
+            // Remove the max-height to allow the element to grow
+            el.style.maxHeight = 'none';
+        },
 
-            this.$nextTick(() => {
-                this.height = body.scrollHeight + 1;
-                this.useInitialHeight = false;
+        beforeLeave(el) {
+            // Restore max-height for the leave transition
+            el.style.maxHeight = el.scrollHeight + 'px';
 
-                if (!this.isOpen) {
-                    body.style.display = 'none';
-                }
-            });
+            // Force repaint
+            el.offsetHeight; // eslint-disable-line no-unused-expressions
+        },
+
+        onLeave(el) {
+            el.style.maxHeight = 0;
+            this.$emit('close');
         }
     },
 
     components: {
         UiIcon,
         UiRippleInk
-    },
-
-    mixins: [
-        RespondsToWindowResize
-    ]
+    }
 };
 </script>
 
@@ -231,8 +216,9 @@ $ui-collapsible-header-background-hover     : $md-grey-300 !default;
 }
 
 .ui-collapsible__body-wrapper {
-    height: initial;
+    max-height: 0;
     overflow: hidden;
+    transition: max-height 0.3s ease;
 }
 
 .ui-collapsible__body {
@@ -241,15 +227,5 @@ $ui-collapsible-header-background-hover     : $md-grey-300 !default;
     display: block;
     padding: rem(16px);
     width: 100%;
-}
-
-.ui-collapsible--transition-toggle-enter-active,
-.ui-collapsible--transition-toggle-leave-active {
-    transition: height 0.3s ease;
-}
-
-.ui-collapsible--transition-toggle-enter,
-.ui-collapsible--transition-toggle-leave-active {
-    height: 0!important;
 }
 </style>